refactor(python): split mitmproxy framing tests
